+++ boot/bootsect.h
@@ -8,4 +8,4 @@ extern "C" uint32_t vbe_info_vector;
 extern "C" void halt(char const*);
 
 uint8_t read_lba_sectors(char *buf, uint8_t drive,
-                         uint32_t lba, uint16_t count);
+                         uint64_t lba, uint16_t count);
